requirements.txt
Django==1.3.0
Django-picklefield==0.1.9
